Restructure pet class
